+++ CHANGELOG.md
@@ -8,11 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The following sections are considered for each release: **Added, Changed, Fixed, Security, Deprecated, Removed**
 
 ## [Unreleased]
+
+### Added
+- On task promotion, detect dangling reference set members and non-defining relationships left by Authoring Platform retire/delete actions and offer to tidy them (delete unreleased, inactivate released) before promoting (#1753)
 - When a task changes to status 'promoted' or 'completed' the associated ticket will now be automatically be changed to state 'Closed'
 - When a task is either deleted or has been released the ticket assocation to this task will now be automatically deleted.
-- The Rows per page: 'All' option on the tasks page now correctly shows 1 - n of n instead of n + 1 - n  of n for the items displayed
 - The display of the tickets backlog page now includes the total number of tickets that is returned for a query.
 
+### Fixed
+- The Rows per page: 'All' option on the tasks page now correctly shows 1 - n of n instead of n + 1 - n  of n for the items displayed
 
 ## [1.3.38] - 2026-04-24